+++ python/tests/plugins/test_primitive_type_checking_impl.py
@@ -1,0 +1,5 @@
+from unittest import TestCase
+
+
+class TestPrimitiveTypeChecking(TestCase):
+    pass
